Fixes for 64bit systems .. mostly cosmetic avoidance of compiler warnings.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@22282 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSData.m
@@ -3758,7 +3758,7 @@ getBytes(void* dst, void* src, unsigned len, unsigned limit, unsigned *pos)
 	    size, GSLastErrorStr(errno)];
 	}
       tmp = shmat(newid, 0, 0);
-      if ((int)tmp == -1)			/* Attached memory? */
+      if ((intptr_t)tmp == -1)			/* Attached memory? */
 	{
 	  [NSException raise: NSMallocException
 		      format: @"Unable to attach to shared memory segment."];
